+++ src/main/java/com/robertx22/uncommon/utilityclasses/HealthUtils.java
@@ -1,6 +1,6 @@
 package com.robertx22.uncommon.utilityclasses;
 
-import com.robertx22.database.stats.types.Health;
+import com.robertx22.database.stats.types.resources.Health;
 import com.robertx22.saveclasses.Unit;
 import com.robertx22.uncommon.datasaving.UnitSaving;
 
@@ -8,7 +8,7 @@ import net.minecraft.entity.EntityLivingBase;
 
 public class HealthUtils {
 
-	public static float DamageToMinecraftHealth(int dmg, EntityLivingBase entity) {
+	public static float DamageToMinecraftHealth(float dmg, EntityLivingBase entity) {
 
 		try {
 			Unit unit = UnitSaving.Load(entity);
